pyproject.toml
[project]
name = "job-search-assistant"
version = "0.1.0"
description = "AI-powered job search assistant using LangGraph agents for resume optimization and job matching"
readme = "README.md"
authors = [
    { name = "georgeliu1998", email = "12277063+georgeliu1998@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.13",
    "langfuse>=2.60.6",
    "langgraph>=0.4.5",
    "pydantic>=2.11.5",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.0",
    "streamlit>=1.45.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Configure hatchling to include all packages from src directory
[tool.hatch.build.targets.wheel]
packages = ["find:"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
